Add start and finish event

+++ nback/Views/Main.swift
@@ -7,10 +7,21 @@
 
 import SwiftUI
 
-enum Types: Character {
-    case Position = "p"
-    case Symbol = "s"
-    case Color = "c"
+func getKeyByType(_ type: String) -> Character {
+    switch type {
+    case "Position":
+        return "p"
+    case "Audio":
+        return "a"
+    case "Color":
+        return "a"
+    case "Shape":
+        return "a"
+    case "Math":
+        return "a"
+    default:
+        return "?"
+    }
 }
 
 struct Elements {
@@ -37,13 +48,14 @@ func nextElements() -> Elements {
 
 
 struct Main: View {
-    @State private var activeTypes: Array<Types> = [.Position, .Color, .Symbol]
-    
+    @Binding var isRunnings: Bool
     @Binding var backgroundColor: Color
-    @Binding var level: Int
-    @Binding var trialTime: Int
-    @Binding var numberOfTrials: Int
-
+    var level: Int
+    var trialTime: Int
+    var numberOfTrials: Int
+    var selectedModes: [String]
+    
+    @State var currentTrial: Int = 0
     @State private var queue: Queue = Queue()
     @State private var displayed: Bool = false
     @State private var symbolColor: Color = Color.white
@@ -53,37 +65,42 @@ struct Main: View {
     
     var body: some View {
         VStack {
+            Text("\(currentTrial) / \(numberOfTrials)")
+                .font(.title2)
             LazyHGrid(rows: Array(repeating: GridItem(.flexible(), spacing: 15), count: 3), spacing: 40) {
                 ForEach(0..<9, id:\.self){index in
                     ZStack{
-                        ForEach(activeTypes, id: \.self) {activeType in
-                            Button("Type X matched") { // TODO: fix
-                                if queue.head  == queue.tail {
-                                    backgroundColor = Color.green.opacity(0.5)
+                        if isRunnings {
+                            ForEach(selectedModes, id: \.self) {selectedMode in
+                                Button("Type X matched") { // TODO: fix
+                                    if queue.head  == queue.tail {
+                                        backgroundColor = Color.green.opacity(0.5)
+                                    } else {
+                                        backgroundColor = Color.red.opacity(0.5)
+                                    }
+                                    
+                                }.keyboardShortcut(KeyEquivalent(getKeyByType(selectedMode)), modifiers: [])
+                            }
+                            
+                            
+                            if let color = elements.color {
+                                if index == elements.position && displayed {
+                                    color
                                 } else {
-                                    backgroundColor = Color.red.opacity(0.5)
+                                    Color.white
                                 }
-                                
-                            }.keyboardShortcut(KeyEquivalent(activeType.rawValue), modifiers: [])
-                        }
-                        
-                        
-                        if let color = elements.color {
-                            if index == elements.position && displayed {
-                                color
-                            } else {
-                                Color.white
                             }
-                        }
-                        
-                        if displayed {
-                            if let symbol = elements.symbol {
-                                Text(elements.position == index ? symbol : "")
-                                    .foregroundColor(symbolColor)
-                                    .font(.system(size: 40))
+                            
+                            if displayed {
+                                if let symbol = elements.symbol {
+                                    Text(elements.position == index ? symbol : "")
+                                        .foregroundColor(symbolColor)
+                                        .font(.system(size: 40))
+                                }
                             }
+                        } else {
+                            Color.white
                         }
-                        
                     }
                     .cornerRadius(10)
                     .frame(width: 120, height: 120)
@@ -92,30 +109,39 @@ struct Main: View {
             .padding(15)
             .frame(width: 500, height: 500)
             .onReceive(timer) { _ in
-                if displayed {
-                    displayed = false
-                    symbolColor = Color.white
-                } else {
-                    displayed = true
-                    symbolColor = Color.black
-                    
-                    elements = nextElements()
-                    
-                    queue.enqueue(elements.position) // TODO:
-                    if queue.size > level + 1 {
-                        queue.drop()
+                if isRunnings {
+                    if displayed {
+                        displayed = false
+                        symbolColor = Color.white
+                        
+                        currentTrial += 1
+                        if currentTrial == numberOfTrials {
+                            isRunnings = false
+                            currentTrial = 0
+                        }
+                    } else {
+                        displayed = true
+                        symbolColor = Color.black
+                        
+                        elements = nextElements()
+                        
+                        queue.enqueue(elements.position) // TODO:
+                        if queue.size > level + 1 {
+                            queue.drop()
+                        }
                     }
+                    
+                    backgroundColor = .black.opacity(0.0)
+                } else {
+                    currentTrial = 0
                 }
-                
-                backgroundColor = .black.opacity(0.0)
             }
-        }
+        }.frame(width: 550, height: 550)
     }
 }
 
 struct Main_Previews: PreviewProvider {
-        static var previews: some View {
-            Main(backgroundColor: .constant(.black), level: .constant(2), trialTime: .constant(1500), numberOfTrials: .constant(25))
-            .frame(width: 500, height: 500)
+    static var previews: some View {
+        Main(isRunnings: .constant(true), backgroundColor: .constant(.black), level: 2, trialTime: 1500, numberOfTrials: 25, selectedModes: ["Position", "Audio"])
     }
 }
